# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Street Cred API
  version: "1.0.0"
  description: |
    Server-authoritative backend for Street Cred (Trapcity).

    **Auth**: send `Authorization: Bearer <accessToken>`. Access tokens last 15 minutes; on
    `401 INVALID_TOKEN` call `/auth/refresh` with the stored refresh token and retry once.
    Refresh tokens rotate: always store the newest one.

    **Idempotency**: for actions that spend or grant currency send `Idempotency-Key: <GUID per user action>`.
    A retry with the same key returns the original response instead of charging twice.

    **Rate limits**: any endpoint can answer `429 RATE_LIMITED`; sign-in and sign-up are limited more tightly.

    **Server is authoritative**: apply the `profile` object returned by every mutating call instead of
    changing balances locally.
servers:
  - url: http://localhost:3000/api/v1
    description: Local development
  - url: https://staging.example.invalid/api/v1
    description: Staging (to be provisioned)
...
security:
  - bearerAuth: []
paths:
...
  /missions/{id}:
    get:
      tags: [Missions]
      summary: One mission
      description: "Details for one mission, e.g. when opening a mission card from a push notification or a saved id."
      parameters: [{ $ref: "#/components/parameters/UuidId" }]
      responses:
        "200":
          description: Mission
          content:
            application/json:
              schema:
                type: object
                properties: { mission: { $ref: "#/components/schemas/Mission" } }
        "404": { $ref: "#/components/responses/Error" }
...
  /admin/users:
    get:
      tags: [Admin]
      summary: Search players by email prefix
      description: "Support tool: find a player by the start of their email address."
      parameters:
        - { name: email, in: query, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 25 } }
      responses:
        "200": { $ref: "#/components/responses/Ok" }
        "403": { $ref: "#/components/responses/Error" }
  /admin/users/{id}:
    get:
      tags: [Admin]
      summary: One player with profile and recent ledger rows
      description: "Support tool: one player with their balances and last 50 ledger rows, for investigating a complaint."
      parameters: [{ $ref: "#/components/parameters/UuidId" }]
      responses:
        "403": { $ref: "#/components/responses/Error", description: "Admin role required" }
        "200": { $ref: "#/components/responses/Ok" }
        "404": { $ref: "#/components/responses/Error" }
  /admin/users/{id}/ban:
    post:
      tags: [Admin]
      summary: Ban a player
      description: "Blocks every request from this account. Existing sessions keep working only until their next request, which is rejected."
      parameters: [{ $ref: "#/components/parameters/UuidId" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties: { reason: { type: string, maxLength: 500 } }
      responses:
        "403": { $ref: "#/components/responses/Error", description: "Admin role required" }
        "204": { description: Banned }
  /admin/users/{id}/unban:
    post:
      tags: [Admin]
      summary: Lift a ban
      description: "Restores access. The player does not need to sign in again."
      parameters: [{ $ref: "#/components/parameters/UuidId" }]
      responses:
        "403": { $ref: "#/components/responses/Error", description: "Admin role required" }
        "204": { description: Unbanned }
  /admin/users/{id}/wallet-adjustments:
    post:
      tags: [Admin]
      summary: Correct a player's balance (recorded in the ledger)
      description: "Give or take cash and gems manually, for refunds and corrections. Always recorded in the ledger with the note."
      parameters:
        - { $ref: "#/components/parameters/UuidId" }
        - { $ref: "#/components/parameters/IdempotencyKeyRequired" }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [note]
              properties:
                cash: { type: integer }
                gems: { type: integer }
                note: { type: string, maxLength: 500 }
      responses:
        "403": { $ref: "#/components/responses/Error", description: "Admin role required" }
        "200":
          description: Adjusted
          content:
            application/json:
              schema:
                type: object
                properties: { profile: { $ref: "#/components/schemas/Profile" } }
        "409": { $ref: "#/components/responses/Error", description: "INSUFFICIENT_FUNDS" }
  /admin/config:
    get:
      tags: [Admin]
      summary: All config rows plus the effective values
      description: "Every tunable with its stored value, plus the values currently in effect."
      responses:
        "403": { $ref: "#/components/responses/Error", description: "Admin role required" }
        "200": { $ref: "#/components/responses/Ok" }
  /admin/config/{key}:
    put:
      tags: [Admin]
      summary: Change one tunable (applies within 30 seconds)
      description: "Tune the game live: prices, XP curve, stamina rate, limits. Values are validated per key and reach all players within 30 seconds."
      parameters:
        - { name: key, in: path, required: true, schema: { type: string }, example: stamina_regen_seconds }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [value]
              properties: { value: { description: Type is validated per key } }
      responses:
        "403": { $ref: "#/components/responses/Error", description: "Admin role required" }
        "200": { $ref: "#/components/responses/Ok" }
        "400": { $ref: "#/components/responses/Error" }
  /admin/missions:
    get:
      tags: [Admin]
      summary: List missions (including inactive ones)
      description: "All missions, including inactive and scheduled ones."
      parameters:
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 500, default: 100 } }
      responses:
        "403": { $ref: "#/components/responses/Error", description: "Admin role required" }
        "200": { $ref: "#/components/responses/Ok" }
    post:
      tags: [Admin]
      summary: Create a mission
      description: "Place a new mission on the map. It appears in nearby searches immediately unless you schedule it with startsAt."
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MissionInput" }
      responses:
        "403": { $ref: "#/components/responses/Error", description: "Admin role required" }
        "201": { $ref: "#/components/responses/Ok" }
        "400": { $ref: "#/components/responses/Error" }
  /admin/missions/{id}:
    patch:
      tags: [Admin]
      summary: Edit a mission (only the fields you send; latitude and longitude go together)
      description: "Change a live mission, e.g. its rewards, or set isActive false to take it off the map. Attempts already running keep the rewards they started with."
      parameters: [{ $ref: "#/components/parameters/UuidId" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MissionInput" }
      responses:
        "403": { $ref: "#/components/responses/Error", description: "Admin role required" }
        "200": { $ref: "#/components/responses/Ok" }
        "404": { $ref: "#/components/responses/Error" }
  /admin/vehicles/{id}:
    put:
      tags: [Admin]
      summary: Create or replace a catalog vehicle (id must match the Unity vehicle id)
      description: "Add or update a vehicle. The id must be the same string the Unity project uses for that vehicle."
      parameters: [{ $ref: "#/components/parameters/CatalogId" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
                description: { type: string }
                priceCash: { type: integer, minimum: 0, default: 0 }
                priceGems: { type: integer, minimum: 0, default: 0 }
                maxUpgradeLevel: { type: integer, default: 5 }
                upgradeBaseCostCash: { type: integer, default: 500 }
                minPlayerLevel: { type: integer, default: 1 }
                isStarter: { type: boolean, default: false }
                isActive: { type: boolean, default: true }
                sortOrder: { type: integer, default: 0 }
      responses:
        "403": { $ref: "#/components/responses/Error", description: "Admin role required" }
        "200": { $ref: "#/components/responses/Ok" }
  /admin/items/{id}:
    put:
      tags: [Admin]
      summary: Create or replace a catalog item
      description: "Add or update an item (outfit, weapon, part or consumable). The id must match the Unity item id."
      parameters: [{ $ref: "#/components/parameters/CatalogId" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, category]
              properties:
                name: { type: string }
                description: { type: string }
                category: { type: string, enum: [Weapon, Outfit, VehiclePart, Consumable] }
                rarity: { type: string, enum: [Common, Rare, Epic, Legendary], default: Common }
                sellValueCash: { type: integer, minimum: 0, default: 0 }
                isStackable: { type: boolean, default: false }
                isActive: { type: boolean, default: true }
      responses:
        "403": { $ref: "#/components/responses/Error", description: "Admin role required" }
        "200": { $ref: "#/components/responses/Ok" }
  /admin/offers/{id}:
    put:
      tags: [Admin]
      summary: Create or replace a store offer
      description: "Put an item on sale, optionally with a purchase limit per player and a start and end time."
      parameters: [{ $ref: "#/components/parameters/CatalogId" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, itemId]
              properties:
                title: { type: string }
                itemId: { type: string }
                quantity: { type: integer, minimum: 1, default: 1 }
                priceCash: { type: integer, minimum: 0, default: 0 }
                priceGems: { type: integer, minimum: 0, default: 0 }
                maxPerUser: { type: [integer, "null"], default: null }
                isActive: { type: boolean, default: true }
                startsAt: { type: [string, "null"], format: date-time }
                endsAt: { type: [string, "null"], format: date-time }
                sortOrder: { type: integer, default: 0 }
      responses:
        "200": { $ref: "#/components/responses/Ok" }
        "403": { $ref: "#/components/responses/Error", description: "Admin role required" }
        "404": { $ref: "#/components/responses/Error", description: "The item the offer sells does not exist" }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    UuidId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
    CatalogId:
      name: id
      in: path
      required: true
      description: Catalog id such as `veh_starter_drifter`
      schema: { type: string, pattern: "^[A-Za-z0-9_.-]+$" }
...
    IdempotencyKeyRequired:
      name: Idempotency-Key
      in: header
      required: true
      description: Required. A GUID per user action; retries with the same key return the first response.
      schema: { type: string, minLength: 8, maxLength: 128 }
  responses:
    Ok:
      description: Success
      content:
        application/json:
          schema: { type: object, additionalProperties: true }
    Error:
      description: Failure
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string, example: INSUFFICIENT_FUNDS }
            message: { type: string }
            details: { description: Optional context, e.g. failing fields or availableAt }
...
    Stamina:
      type: object
      properties:
        current: { type: integer }
        max: { type: integer }
        regenSeconds: { type: integer }
        nextPointAt: { type: [string, "null"], format: date-time }
        fullAt: { type: [string, "null"], format: date-time }
    Profile:
      type: object
      description: The authoritative player state. Replace local values with this on every response.
      properties:
        level: { type: integer }
        xp: { type: integer, description: Progress inside the current level }
        xpToNextLevel: { type: integer }
        totalXp: { type: integer }
        cash: { type: integer }
        gems: { type: integer }
        stamina: { $ref: "#/components/schemas/Stamina" }
        equippedVehicleId: { type: [string, "null"] }
        equippedOutfitId: { type: [string, "null"] }
        hasFoundersPass: { type: boolean }
...
    Reward:
      type: object
      properties:
        cash: { type: integer }
        gems: { type: integer }
        xp: { type: integer }
        itemId: { type: [string, "null"] }
    Mission:
      type: object
      properties:
        id: { type: string, format: uuid }
        title: { type: string }
        description: { type: string }
        missionType: { type: string, enum: [DeadDropCourier, TurfSkirmishAR, ZoneReconWalking, HighRollerSprint] }
        latitude: { type: number }
        longitude: { type: number }
        triggerRadiusM: { type: number }
        timeLimitS: { type: integer }
        requiredStamina: { type: integer }
        reward: { $ref: "#/components/schemas/Reward" }
        minPlayerLevel: { type: integer }
        cooldownS: { type: [integer, "null"], description: null means the mission can only be completed once }
...
    MissionInput:
      type: object
      description: POST requires title, missionType, latitude, longitude. PATCH accepts any subset.
      properties:
        title: { type: string }
        description: { type: string }
        missionType: { type: string, enum: [DeadDropCourier, TurfSkirmishAR, ZoneReconWalking, HighRollerSprint] }
        latitude: { type: number, minimum: -90, maximum: 90 }
        longitude: { type: number, minimum: -180, maximum: 180 }
        triggerRadiusM: { type: number, default: 30 }
        timeLimitS: { type: integer, default: 120 }
        requiredStamina: { type: integer, default: 15 }
        rewardCash: { type: integer, default: 500 }
        rewardGems: { type: integer, default: 5 }
        rewardXp: { type: integer, default: 250 }
        rewardItemId: { type: [string, "null"] }
        minPlayerLevel: { type: integer, default: 1 }
        cooldownS: { type: [integer, "null"], default: 3600 }
        minCompletionS: { type: integer, default: 10 }
        isActive: { type: boolean, default: true }
        startsAt: { type: [string, "null"], format: date-time }
        endsAt: { type: [string, "null"], format: date-time }